requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paddlepaddle==2.6.0
paddleocr==2.7.3
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.2.0
numpy<2.0.0
